Apply changes as reviewer suggests
