+++ packages/tools/src/stateManagement/segmentation/utilities/getAnnotationsUIDMapFromSegmentation.ts
@@ -1,0 +1,28 @@
+import type { ContourSegmentationData } from '../../../types';
+import { getSegmentation } from '../getSegmentation';
+
+/**
+ * Retrieves the annotationUIDsMap for a given segmentation ID, if available.
+ *
+ * @param segmentationId - The segmentation ID
+ * @returns The annotationUIDsMap (Map<segmentIndex, Set<annotationUID>>) or undefined
+ */
+export function getAnnotationsUIDMapFromSegmentation(segmentationId: string) {
+  const segmentation = getSegmentation(segmentationId);
+  if (!segmentation) {
+    return;
+  }
+
+  const contourRepresentationData = segmentation.representationData
+    ?.Contour as ContourSegmentationData;
+
+  if (!contourRepresentationData) {
+    return;
+  }
+
+  const { annotationUIDsMap } = contourRepresentationData;
+  if (!annotationUIDsMap) {
+    return;
+  }
+  return annotationUIDsMap;
+}
